+++ psfws/__init__.py
@@ -2,5 +2,4 @@ __author__ = "Claire-Alice Hebert"
 __email__ = "chebert@stanford.edu"
 __version__ = "0.1"
 
-from . import gen_params
-from . import utils
+from .gen_params import ParameterGenerator
